Add phones and words to db
